PMS-23 Erstellung der DTO Klasse zum Lesen aller Projekte eines Kunden/verantwortlichen Mitarbeiters
- DTO: EmployeeIds von integer zu Long geändert

+++ src/main/java/de/szut/lf8_project/project/dto/GetProjectDto.java
@@ -21,5 +21,5 @@ public class GetProjectDto {
 
     private LocalDateTime endDate;
 
-    private Set<Integer> employeeIds;
+    private Set<Long> employeeIds;
 }
